refactor(front): Rename ImageInspector to Inspector
