+++ ui/components/datasetItemWorkspace/src/components/Inspector/ArrowSvg.svelte
