+++ ui/components/datasetItemWorkspace/src/components/Inspector/ObjectCard.svelte
