+++ ui/components/datasetItemWorkspace/src/components/Inspector/ObjectsInspector.svelte
