+++ ui/components/datasetItemWorkspace/src/components/Inspector/ObjectsModelSection.svelte
