+++ ui/components/datasetItemWorkspace/src/components/Inspector/SceneInspector.svelte
